requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0
sentence-transformers==2.3.1
faiss-cpu==1.13.2
requests==2.31.0
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2
tenacity==8.2.3
